Changed allocators to use size_t for memory sizes

+++ include/core/allocator.h
@@ -16,9 +16,9 @@ namespace fissura
 		Allocator();
 		virtual ~Allocator();
 
-		virtual void* allocate(u32 size, u8 alignment) = 0;
+		virtual void* allocate(size_t size, u8 alignment) = 0;
 		virtual void deallocate(void* p) = 0;
-		virtual u32 getTotalUsedMemory() const = 0;
+		virtual size_t getTotalUsedMemory() const = 0;
 		virtual u32 getTotalNumAllocations() const = 0;
 		virtual bool canDeallocate() const = 0;
